Apply hlum's improvement tasks from code review
- Create ShortcutViewModel: move shortcut-related state/methods from SettingsViewModel
  (currentHotkey, actionType, isRecording, isSuccessState, tempModifiers, tempKeyDisplay,
  startRecording, stopRecording, completeRecording, saveHotkey, saveActionType, runTestShortcut)
- SettingsViewModel: remove shortcut-related code (separation of concerns)
- ShortcutView: use @StateObject private var viewModel = ShortcutViewModel()
- ShortcutView: fix Binding warning - replace Binding(get:set:) with $viewModel.actionType + .onChange
- SettingsView: remove ShortcutView's dependency on SettingsViewModel
- SettingView: fix window focus bug - move makeKeyAndOrderFront to .onAppear so it runs on every open
- Create NSEvent+ModifierFlags extension with formattedString property
- Hotkey.displayString: use modifiers.formattedString extension
- RecorderOverlaySectionView: change let to @Binding for mutable state properties
- RecorderOverlaySectionView: use tempModifiers.formattedString instead of KeySender.formatModifiers
- ShortcutView_Previews: rename from RaycastSettingsView_Previews
