+++ yt-mac-menu/yt-mac-menu/Domain/Models/NSEvent+ModifierFlags.swift
@@ -1,0 +1,12 @@
+import AppKit
+
+extension NSEvent.ModifierFlags {
+    var formattedString: String {
+        var keys: [String] = []
+        if contains(.control) { keys.append("⌃") }
+        if contains(.option)  { keys.append("⌥") }
+        if contains(.shift)   { keys.append("⇧") }
+        if contains(.command) { keys.append("⌘") }
+        return keys.joined(separator: " ")
+    }
+}
